Move feedbackWrapper logic to FeedbackForm
