+++ .github/workflows/azure-static-web-apps-jolly-sand-0ac78c710.yml
@@ -57,6 +57,7 @@ jobs:
       - name: Build And Deploy
         env:
           REACT_APP_CLIENT_ID: ${{ secrets.REACT_APP_CLIENT_ID }}
+          REACT_APP_FEEDBACK_CAMPAIGN_ID: ${{ secrets.REACT_APP_FEEDBACK_CAMPAIGN_ID }}
         id: builddeploy
         uses: Azure/static-web-apps-deploy@v1
         with:
